project_one/requirements.txt
fastapi
fastapi[standard]
httpx
py-automapper
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-jaeger